+++ packages/dialtone-vue3/components/rich_text_editor/rich_text_editor.test.js
@@ -108,6 +108,9 @@ describe('DtRichTextEditor tests', () => {
             type: 'doc',
             content: [{
               type: 'paragraph',
+              attrs: {
+                textAlign: 'left',
+              },
               content: [{
                 text: 'new value',
                 type: 'text',